Separate helper functions into own file to resolve circular imports
